# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Geolocal",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "66068697"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign up",
                "description": "",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password",
                                    "firstname",
                                    "lastname",
                                    "communes",
                                    "genre",
                                    "rule"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Damas"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Helas"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "66068697"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "commune_id": {
                                        "type": "int",
                                        "format": "int",
                                        "example": "1"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "M"
                                    },
                                    "rule": {
                                        "type": "string",
                                        "format": "admin,admin_locaux,users",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "info sur l'utilisateur",
                "description": "",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Vous devez vous connectez d'abord",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "liste des utilisateurs",
                "description": "",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Vous devez vous connectez d'abord et etre admin ou admin_locaux",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "info sur un utilisateur seulement un admin et admin_locaux",
                "description": "",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Vous devez vous connectez d'abord",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "suprimer un utilisateur",
                "description": "",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Vous devez utiliser le token de l'utilisateur ou un admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "users"
                ],
                "summary": "mise a jour des infos de l'utilisateurs",
                "description": "",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Vous devez utiliser le token de l'utilisateur ou un admin ou admin_locaux",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "password",
                                    "firstname",
                                    "lastname",
                                    "communes",
                                    "genre",
                                    "rule"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Damas"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Helas"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "66068697"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "commune_id": {
                                        "type": "int",
                                        "format": "int",
                                        "example": "1"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "M"
                                    },
                                    "rule": {
                                        "type": "string",
                                        "format": "admin,admin_locaux,users",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}